Init DataBase User(admin/admin)
Admin role with email created

+++ src/main/java/simSalProject/business/UserBusiness.java
@@ -7,9 +7,9 @@ import java.util.List;
 import javax.faces.bean.RequestScoped;
 import javax.inject.Inject;
 import javax.inject.Named;
-import javax.ws.rs.core.Response;
 
 import simSalProject.models.User_;
+import simSalProject.models.User_.UserRole;
 import simSalProject.repositories.UserRepository;
 
 @Named("UserBus")
@@ -22,8 +22,19 @@ public class UserBusiness  {
 	UserRepository USER_DB;
 	
 	
-	public void createUser(User_ myUser) {
-		USER_DB.createEntity(myUser);
+	public String createUser(User_ myUser) {
+		if(myUser.getUserRole() == UserRole.USER) {
+			USER_DB.createEntity(myUser);	
+			
+		} else if (myUser.getUserRole() == UserRole.ADMIN) {
+			return "Only 1 admin can exist";
+		}
+		return "User created";
+	}
+	
+	public String createAdmin(User_ adminUser) {
+		USER_DB.createEntity(adminUser);
+		return "ADMIN didn't exist, ADMIN created with default credentials";
 	}
 	
 	
@@ -33,11 +44,15 @@ public class UserBusiness  {
 	}
 	
 	public void editUser(long id, User_ myUserToEdit) {
-		USER_DB.editEntity(myUserToEdit);
+		if (myUserToEdit.getUserRole() != UserRole.ADMIN) {
+			USER_DB.editEntity(myUserToEdit);
+		}
 	}
 	
 	public void removeUser(User_ myUser) {
-		USER_DB.removeEntity(myUser);
+		if (myUser.getUserRole() != UserRole.ADMIN) {
+			USER_DB.removeEntity(myUser);
+		}
 	}
 	
 	public List<Long> getAllIds() {
@@ -48,13 +63,28 @@ public class UserBusiness  {
 		return USER_DB.allValues();
 	}
 	
+	public String initDataBase() {
+		String message = "";
+		//0 is ENUM(0), which is ADMIN
+		if (USER_DB.getRoleCount(UserRole.ADMIN) == 0){
+			User_ initUser = new User_();
+			initUser.setEmail("admin@admin.com");
+			initUser.setPassword("admin");
+			initUser.setUserRole(UserRole.ADMIN);
+			message = createAdmin(initUser);
+			
+		} else {
+			message = "ADMIN already exists";
+		}
+		
+		return message;
+	}
 	
-	
-	public Response login() {
+	public String login() {
 		
 		
 		
-		return Response.ok().build();
+		return "";
 		
 	}
